make summary tiles consistent across
